add --start option and add fallback for rpc port occupied
