+++ iroh/src/commands/repl.rs
@@ -1,8 +1,9 @@
 use anyhow::Result;
 use clap::{Parser, Subcommand};
 use colored::Colorize;
-use iroh::client::quic::Iroh;
+use iroh::{client::Iroh, rpc_protocol::ProviderService};
 use iroh_base::base32::fmt_short;
+use quic_rpc::ServiceConnection;
 use rustyline::{error::ReadlineError, Config, DefaultEditor};
 use tokio::sync::{mpsc, oneshot};
 
@@ -11,7 +12,10 @@ use crate::{
     config::{ConsoleEnv, ConsolePaths},
 };
 
-pub async fn run(iroh: &Iroh, env: &ConsoleEnv) -> Result<()> {
+pub async fn run<C>(iroh: &Iroh<C>, env: &ConsoleEnv) -> Result<()>
+where
+    C: ServiceConnection<ProviderService>,
+{
     println!("{}", "Welcome to the Iroh console!".purple().bold());
     println!("Type `{}` for a list of commands.", "help".bold());
     let mut from_repl = Repl::spawn(env.clone());
